requirements.txt
# Core LlamaIndex dependencies
llama-index>=0.9.0
llama-index-llms-openai>=0.1.0
llama-index-utils-workflow>=0.1.0
llama-index-embeddings-openai>=0.1.0
llama-index-vector-stores-faiss>=0.1.0

# Vector databases
faiss-cpu>=1.7.4
chromadb>=0.4.0

# API clients
tavily-python>=0.3.0
sec-api>=1.0.0
aiohttp>=3.8.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
pydantic>=2.0.0

# Financial data
yfinance>=0.2.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.0
PyYAML>=6.0
rich>=13.0.0

# Embeddings (local and OpenAI)
sentence-transformers>=2.2.0
scikit-learn>=1.3.0

# Sentiment analysis (optional)
transformers>=4.30.0
torch>=2.0.0

# Development
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
